refactor: Move flattenText function
